[5.x] Merge develop (#1369)
* [develop] Adds L11 support (#1352)

* Adds L11 support

* Fixes service provider installation

* Fixes service provider installation

* Fixes service provider installation

* Fixes static analysis

* Improves test suite organization

* Test without `runInSeparateProcess`

* Adjusts installation process (#1353)

+++ src/Console/InstallCommand.php
@@ -3,6 +3,7 @@
 namespace Laravel\Horizon\Console;
 
 use Illuminate\Console\Command;
+use Illuminate\Support\ServiceProvider;
 use Illuminate\Support\Str;
 
 class InstallCommand extends Command
@@ -51,17 +52,22 @@ class InstallCommand extends Command
     {
         $namespace = Str::replaceLast('\\', '', $this->laravel->getNamespace());
 
-        $appConfig = file_get_contents(config_path('app.php'));
+        if (file_exists($this->laravel->bootstrapPath('providers.php'))) {
+            // @phpstan-ignore-next-line
+            ServiceProvider::addProviderToBootstrapFile("{$namespace}\\Providers\\HorizonServiceProvider");
+        } else {
+            $appConfig = file_get_contents(config_path('app.php'));
 
-        if (Str::contains($appConfig, $namespace.'\\Providers\\HorizonServiceProvider::class')) {
-            return;
+            if (Str::contains($appConfig, $namespace.'\\Providers\\HorizonServiceProvider::class')) {
+                return;
+            }
+
+            file_put_contents(config_path('app.php'), str_replace(
+                "{$namespace}\\Providers\EventServiceProvider::class,".PHP_EOL,
+                "{$namespace}\\Providers\EventServiceProvider::class,".PHP_EOL."        {$namespace}\Providers\HorizonServiceProvider::class,".PHP_EOL,
+                $appConfig
+            ));
         }
-
-        file_put_contents(config_path('app.php'), str_replace(
-            "{$namespace}\\Providers\EventServiceProvider::class,".PHP_EOL,
-            "{$namespace}\\Providers\EventServiceProvider::class,".PHP_EOL."        {$namespace}\Providers\HorizonServiceProvider::class,".PHP_EOL,
-            $appConfig
-        ));
 
         file_put_contents(app_path('Providers/HorizonServiceProvider.php'), str_replace(
             "namespace App\Providers;",
